Feat/swarm plum pork (#2343)
* Moved Swarm FindPeers from node API to plum pork
* Moved Swarm Connect from node API to plum pork
* Moved Swarm Peer from node API to plum pork
* Removed unnecessary SwarmConnectResult type
* Rename SwarmFindPeer to DhtFindPeer, fix broken miner_test test due to empty pubkey
* Left comment w/ issue link for dht findprovs test
