--- api/swarm.go
@@ -1,64 +1,0 @@
-package api
-
-import (
-	"context"
-	peerstore "gx/ipfs/QmRhFARzTHcFh8wUxwN5KvyTGq73FLC65EfFAhz8Ng7aGb/go-libp2p-peerstore"
-	peer "gx/ipfs/QmTu65MVbemtUxJEWgsTtzv9Zv9P8rvmqNA4eG9TrTRGYc/go-libp2p-peer"
-)
-
-// Swarm is the interface that defines methods to interact with the p2p swarm of the node.
-type Swarm interface {
-	Peers(ctx context.Context, verbose, latency, streams bool) (*SwarmConnInfos, error)
-	Connect(ctx context.Context, addrs []string) ([]SwarmConnectResult, error)
-	FindPeer(ctx context.Context, peerID peer.ID) (peerstore.PeerInfo, error)
-}
-
-// SwarmConnInfo represents details about a single swarm connection.
-type SwarmConnInfo struct {
-	Addr    string
-	Peer    string
-	Latency string
-	Muxer   string
-	Streams []SwarmStreamInfo
-}
-
-// SwarmStreamInfo represents details about a single swarm stream.
-type SwarmStreamInfo struct {
-	Protocol string
-}
-
-func (ci *SwarmConnInfo) Less(i, j int) bool {
-	return ci.Streams[i].Protocol < ci.Streams[j].Protocol
-}
-
-func (ci *SwarmConnInfo) Len() int {
-	return len(ci.Streams)
-}
-
-func (ci *SwarmConnInfo) Swap(i, j int) {
-	ci.Streams[i], ci.Streams[j] = ci.Streams[j], ci.Streams[i]
-}
-
-// SwarmConnInfos represent details about a list of swarm connections.
-type SwarmConnInfos struct {
-	Peers []SwarmConnInfo
-}
-
-func (ci SwarmConnInfos) Less(i, j int) bool {
-	return ci.Peers[i].Addr < ci.Peers[j].Addr
-}
-
-func (ci SwarmConnInfos) Len() int {
-	return len(ci.Peers)
-}
-
-func (ci SwarmConnInfos) Swap(i, j int) {
-	ci.Peers[i], ci.Peers[j] = ci.Peers[j], ci.Peers[i]
-}
-
-// SwarmConnectResult represents the data generated when trying to connect to another p2p
-// network member.
-type SwarmConnectResult struct {
-	Peer    string
-	Success bool
-}
