+++ net/router.go
@@ -5,6 +5,7 @@ import (
 
 	"gx/ipfs/QmR8BauakNcBa3RbE4nbQu76PDiJgoQgz8AJdhJuiU4TAw/go-cid"
 	pstore "gx/ipfs/QmRhFARzTHcFh8wUxwN5KvyTGq73FLC65EfFAhz8Ng7aGb/go-libp2p-peerstore"
+	"gx/ipfs/QmTu65MVbemtUxJEWgsTtzv9Zv9P8rvmqNA4eG9TrTRGYc/go-libp2p-peer"
 	routing "gx/ipfs/QmWaDSNoSdSXU9b6udyaq9T8y6LkzMwqWxECznFqvtcTsk/go-libp2p-routing"
 )
 
@@ -38,4 +39,9 @@ func NewRouter(r routing.IpfsRouting) *Router {
 // given key.
 func (r *Router) FindProvidersAsync(ctx context.Context, key cid.Cid, count int) <-chan pstore.PeerInfo {
 	return r.routing.FindProvidersAsync(ctx, key, count)
+}
+
+// FindPeer searches the libp2p router for a given peer id
+func (r *Router) FindPeer(ctx context.Context, peerID peer.ID) (pstore.PeerInfo, error) {
+	return r.routing.FindPeer(ctx, peerID)
 }
