+++ tools/fast/actions_swarm.go
@@ -7,12 +7,12 @@ import (
 	"gx/ipfs/QmNTCey11oxhb1AxDnQBRHtdhap6Ctud872NjAYPYYXPuc/go-multiaddr"
 	"gx/ipfs/QmTu65MVbemtUxJEWgsTtzv9Zv9P8rvmqNA4eG9TrTRGYc/go-libp2p-peer"
 
-	"github.com/filecoin-project/go-filecoin/api"
+	"github.com/filecoin-project/go-filecoin/net"
 )
 
 // SwarmConnect runs the `swarm connect` command against the filecoin process
-func (f *Filecoin) SwarmConnect(ctx context.Context, addrs ...multiaddr.Multiaddr) ([]api.SwarmConnectResult, error) {
-	var out []api.SwarmConnectResult
+func (f *Filecoin) SwarmConnect(ctx context.Context, addrs ...multiaddr.Multiaddr) ([]peer.ID, error) {
+	var out []peer.ID
 
 	args := []string{"go-filecoin", "swarm", "connect"}
 
@@ -27,9 +27,9 @@ func (f *Filecoin) SwarmConnect(ctx context.Context, addrs ...multiaddr.Multiadd
 	return out, nil
 }
 
-// SwarmFindpeer runs the `swarm findpeer` command against the filecoin process
-func (f *Filecoin) SwarmFindpeer(ctx context.Context, pid peer.ID) ([]multiaddr.Multiaddr, error) {
-	decoder, err := f.RunCmdLDJSONWithStdin(ctx, nil, "go-filecoin", "swarm", "findpeer", pid.String())
+// DhtFindpeer runs the `dht findpeer` command against the filecoin process
+func (f *Filecoin) DhtFindpeer(ctx context.Context, pid peer.ID) ([]multiaddr.Multiaddr, error) {
+	decoder, err := f.RunCmdLDJSONWithStdin(ctx, nil, "go-filecoin", "dht", "findpeer", pid.String())
 	if err != nil {
 		return nil, err
 	}
@@ -57,8 +57,8 @@ func (f *Filecoin) SwarmFindpeer(ctx context.Context, pid peer.ID) ([]multiaddr.
 }
 
 // SwarmPeers runs the `swarm peers` command against the filecoin process
-func (f *Filecoin) SwarmPeers(ctx context.Context, options ...ActionOption) ([]api.SwarmConnInfo, error) {
-	var out api.SwarmConnInfos
+func (f *Filecoin) SwarmPeers(ctx context.Context, options ...ActionOption) ([]net.SwarmConnInfo, error) {
+	var out net.SwarmConnInfos
 
 	args := []string{"go-filecoin", "swarm", "peers"}
